advance-agent-examples/sub-agents/pyproject.toml
[project]
name = "mavc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.21.0",
    "google-auth>=2.45.0",
    "google-cloud-storage>=3.4.1",
    "uagents-adapter[a2a-outbound]==0.6.2",
]
